Refactor: AuthApi 리팩토링
- 새로운 버전의 OAuth2 로그인 추가
- 기존 버전의 로그인 api는 새로운 버전의 주소로 리다이렉트하도록 리팩토링
