+++ src/main/java/podo/odeego/web/api/auth/v1/AuthApi.java
@@ -1,15 +1,17 @@
-package podo.odeego.web.api.auth;
+package podo.odeego.web.api.auth.v1;
+
+import java.io.IOException;
 
 import javax.servlet.http.HttpServletRequest;
+import javax.servlet.http.HttpServletResponse;
 
+import org.springframework.http.HttpHeaders;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.CookieValue;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
-import podo.odeego.web.api.auth.dto.MemberLoginResponse;
-import podo.odeego.web.auth.MemberLoginManager;
 import podo.odeego.web.auth.dto.ReissueResponse;
 import podo.odeego.web.auth.service.AuthService;
 
@@ -17,21 +19,17 @@ import podo.odeego.web.auth.service.AuthService;
 @RequestMapping("/api/v1/auth")
 public class AuthApi {
 
-	private static final String AUTHORIZATION = "Authorization";
-
-	private final MemberLoginManager memberLoginManager;
-
 	private final AuthService authService;
 
-	public AuthApi(MemberLoginManager memberLoginManager, AuthService authService) {
-		this.memberLoginManager = memberLoginManager;
+	public AuthApi(AuthService authService) {
 		this.authService = authService;
 	}
 
 	@PostMapping("/user/me")
-	public ResponseEntity<MemberLoginResponse> login(HttpServletRequest request) {
-		String oAuth2Token = request.getHeader(AUTHORIZATION);
-		return ResponseEntity.ok(memberLoginManager.login(oAuth2Token));
+	public void login(HttpServletRequest request, HttpServletResponse response) throws
+		IOException {
+		response.setHeader(HttpHeaders.AUTHORIZATION, request.getHeader(HttpHeaders.AUTHORIZATION));
+		response.sendRedirect("/api/v2/auth/login/oauth2");
 	}
 
 	@PostMapping("/reissue")
